backend/requirements.txt
# FastAPI 后端依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis 任务队列
redis==5.0.1
celery==5.3.6

# AI/LLM
openai==1.12.0
anthropic==0.18.0
httpx==0.26.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# 工具库
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 腾讯云SDK
cos-python-sdk-v5==1.9.30
tencentcloud-sdk-python==3.0.1000

# 日志
loguru==0.7.2
pycryptodome>=3.19.0
